chore: move `AIModeration` out of the stack handler
